src/main/java/br/com/alura/escola/academico/aplicacao/aluno/matricular/MatriculadorDeAluno.java
package br.com.alura.escola.academico.aplicacao.aluno.matricular;

import br.com.alura.escola.shared.dominio.evento.PublicadorDeEventos;
import br.com.alura.escola.academico.dominio.aluno.Aluno;
import br.com.alura.escola.academico.dominio.aluno.AlunoMatriculado;
import br.com.alura.escola.academico.dominio.aluno.RepositorioDeAlunos;

public class MatriculadorDeAluno {

    private final RepositorioDeAlunos repositorio;
    private final PublicadorDeEventos publicador;

    public MatriculadorDeAluno(RepositorioDeAlunos repositorio, PublicadorDeEventos publicador) {
        this.repositorio = repositorio;
        this.publicador = publicador;
    }

    public void matricular(DadosDoAlunoDTO dados) {
        Aluno novo = dados.criarAluno();
        repositorio.matricular(novo);
        publicador.publicar(new AlunoMatriculado(novo.getCpf()));
    }
}
